fix https://gitee.com/best_handsome/mybatis-plus-join/issues/I6J2YN

+++ mybatis-plus-join-core/src/main/java/com/github/yulichang/method/SelectJoinCount.java
@@ -28,6 +28,6 @@ public class SelectJoinCount extends MPJAbstractMethod {
         String sql = String.format(sqlMethod.getSql(), sqlFirst(), sqlCount(),
                 tableInfo.getTableName(), sqlAlias(), sqlFrom(), sqlWhereEntityWrapper(true, tableInfo), sqlComment());
         SqlSource sqlSource = languageDriver.createSqlSource(configuration, sql, modelClass);
-        return this.addSelectMappedStatementForOther(mapperClass, sqlMethod.getMethod(), sqlSource, Integer.class);
+        return this.addSelectMappedStatementForOther(mapperClass, sqlMethod.getMethod(), sqlSource, Long.class);
     }
 }
